requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.1.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
disutils==1.4.32.post2
frozenlist==1.5.0
idna==3.10
lxml==5.3.0
multidict==6.1.0
numpy==1.17.5
pandas==1.3.5
pika==1.3.2
propcache==0.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.15.2
